Rename LiveAudioTranscriptionClient to LiveAudioSession and LiveAudioTranscriptionTypes to LiveAudioTypes
