+++ sdk/cs/src/OpenAI/LiveAudioTypes.cs
