+++ sdk/js/src/detail/modelVariant.ts
@@ -4,7 +4,7 @@ import { ModelInfo } from '../types.js';
 import { ChatClient } from '../openai/chatClient.js';
 import { AudioClient } from '../openai/audioClient.js';
 import { EmbeddingClient } from '../openai/embeddingClient.js';
-import { LiveAudioTranscriptionSession } from '../openai/liveAudioTranscriptionClient.js';
+import { LiveAudioTranscriptionSession } from '../openai/liveAudioSession.js';
 import { ResponsesClient } from '../openai/responsesClient.js';
 import { IModel } from '../imodel.js';
