+++ sdk/js/src/imodel.ts
@@ -1,7 +1,7 @@
 import { ChatClient } from './openai/chatClient.js';
 import { AudioClient } from './openai/audioClient.js';
 import { EmbeddingClient } from './openai/embeddingClient.js';
-import { LiveAudioTranscriptionSession } from './openai/liveAudioTranscriptionClient.js';
+import { LiveAudioTranscriptionSession } from './openai/liveAudioSession.js';
 import { ResponsesClient } from './openai/responsesClient.js';
 import { ModelInfo } from './types.js';
