+++ sdk/python/src/openai/audio_client.py
@@ -14,7 +14,7 @@ from typing import Generator, List, Optional
 
 from ..detail.core_interop import CoreInterop, InteropRequest
 from ..exception import FoundryLocalException
-from .live_audio_transcription_client import LiveAudioTranscriptionSession
+from .live_audio_session import LiveAudioTranscriptionSession
 
 logger = logging.getLogger(__name__)
