+++ sdk/rust/src/openai/mod.rs
@@ -2,7 +2,7 @@ mod audio_client;
 mod chat_client;
 mod embedding_client;
 mod json_stream;
-mod live_audio_client;
+mod live_audio_session;
 
 pub use self::audio_client::{
     AudioClient, AudioClientSettings, AudioTranscriptionResponse, AudioTranscriptionStream,
@@ -11,7 +11,7 @@ pub use self::audio_client::{
 pub use self::chat_client::{ChatClient, ChatClientSettings, ChatCompletionStream};
 pub use self::embedding_client::EmbeddingClient;
 pub use self::json_stream::JsonStream;
-pub use self::live_audio_client::{
+pub use self::live_audio_session::{
     ContentPart, CoreErrorResponse, LiveAudioTranscriptionOptions, LiveAudioTranscriptionResponse,
     LiveAudioTranscriptionSession, LiveAudioTranscriptionStream,
 };
